requirements.txt
pyenchant==3.2.2
pyTelegramBotAPI==4.8.0
requests~=2.22.0
translate==3.6.1
wikipedia==1.4.0
aiohttp==3.8.3
gevent~=22.10.2